fix bag cost in check

+++ backend/src/main/java/ru/nsu/sber_portal/ccfit/models/dto/orderDto/OrderPattern.java
@@ -14,5 +14,5 @@ public class OrderPattern {
 
     protected Integer numberTable;
 
-    protected Integer count;
+    protected Long count;
 }
